Consider less lets as interesting
if there is no eta-expansion possible, do not consider that. This is a
win because uninteresting lets can cause the code to choose the wrong
side in a case/app. But it could also be a loss if the interesting call
path passes through a fully expanded function. We'll see.